add validate_data_point to validation
